requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioice==0.10.2
aiortc==1.14.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
audioread==3.1.0
av==16.1.0
babel==2.17.0
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
colorlog==6.10.1
cryptography==46.0.3
csvw==3.7.0
decorator==5.2.1
discord-protos==0.0.2
discord.py-self==2.0.1
dlinfo==2.0.0
dnspython==2.8.0
dotenv==0.9.9
espeakng-loader==0.2.4
fastapi==0.128.0
fastrtc==0.0.34
ffmpy==1.0.0
filelock==3.20.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.1.0
google-crc32c==1.8.0
gradio==5.50.0
gradio_client==1.14.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.2
humanfriendly==10.0
idna==3.11
ifaddr==0.2.0
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kokoro-onnx==0.4.9
language-tags==1.2.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
numba==0.63.1
numpy==2.3.5
onnxruntime==1.23.2
orjson==3.11.5
packaging==25.0
pandas==2.3.3
phonemizer-fork==3.3.2
pillow==11.3.0
platformdirs==4.5.1
pooch==1.8.2
propcache==0.4.1
protobuf==6.33.4
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
pylibsrtp==1.0.0
PyNaCl==1.5.0
pyOpenSSL==25.3.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
rdflib==7.5.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rfc3986==1.5.0
rich==14.2.0
rpds-py==0.30.0
ruff==0.14.13
safehttpx==0.1.7
scikit-learn==1.8.0
scipy==1.17.0
segments==2.3.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
starlette==0.50.0
sympy==1.14.0
termcolor==3.3.0
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
websockets==15.0.1
yarl==1.22.0
